added main method that can be used to test everything using example file or own file

+++ opendut-edgar/restbus-simulation/src/restbus_simulation.rs
@@ -32,8 +32,8 @@ impl RestbusSimulation {
             can_frames.push(
                 create_can_frame_structure(timed_can_frame.can_id, timed_can_frame.len, timed_can_frame.addressing_mode, timed_can_frame.frame_tx_behavior, &timed_can_frame.data_vector));
         
-            create_bcm_structure_bytes(timed_can_frame.count, timed_can_frame.ival1.tv_sec, timed_can_frame.ival1.tv_usec, 
-                timed_can_frame.ival2.tv_sec, timed_can_frame.ival2.tv_usec, timed_can_frame.can_id, timed_can_frame.frame_tx_behavior, &can_frames, &mut write_bytes);
+            create_bcm_structure_bytes(timed_can_frame.count, timed_can_frame.ival1.tv_sec as u64, timed_can_frame.ival1.tv_usec as u64, 
+                timed_can_frame.ival2.tv_sec as u64, timed_can_frame.ival2.tv_usec as u64, timed_can_frame.can_id, timed_can_frame.frame_tx_behavior, &can_frames, &mut write_bytes);
             
             write_bytes_global.push(write_bytes);
